fix(host): umount loop device path before detaching it
